refactor: move app state code to custom hook
